Rename aux.h for windows compatibility
