+++ include/photospline/detail/auxiliary.h
